% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_distance.R
\name{decompose_distance}
\alias{decompose_distance}
\title{Decompose the distance between observed and null}
\usage{
decompose_distance(beam.stats, z = TRUE, mess.freq = 25)
}
\arguments{
\item{beam.stats}{A beam.stats object from compute_beam_stats function}

\item{z}{Logical indicating whether to z-scale each vector of one coefficient estimate across bootstraps before analysis}

\item{mess.freq}{Message frequency; default 25.}
}
\value{
A list of dataframes. List names correspond to set names. Each data.frame has Association column with feature name and association type and Contribution column with relative contribution, ordered greatest to least contribution.
}
\description{
Use distance decomposition to identify relative contribution of each omic feature/endpoint associaiton
}
\examples{
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=10, seed=123)
specs <- prep_beam_specs(beam.data=test.beam.data, endpts=c("MRD29", "EFS", "OS"), firth=TRUE)
test.beam.stats <- compute_beam_stats(beam.data=test.beam.data, beam.specs=specs)
test.dist <- decompose_distance(test.beam.stats, z=TRUE)
}
