% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beam_clin.R
\name{plot_beam_clin}
\alias{plot_beam_clin}
\title{Plot BEAM Sets}
\usage{
plot_beam_clin(
  beam.result,
  beam.specs = NULL,
  beam.set.pvals,
  beam.feat.pvals,
  set.id,
  pair.type = NULL,
  number.pairs = 1,
  n.col = NULL,
  n.row = NULL,
  title.size = 10
)
}
\arguments{
\item{beam.result}{A beam.stats object from compute_beam_stats}

\item{beam.specs}{A data.frame. Default NULL, in which case beam.result$beam.specs is used. Otherwise can input other beam.specs data.frame that must contain name, mtx, mdl, plot columns.}

\item{beam.set.pvals}{A list containing BEAM set p-values from compute_set_pvalues.}

\item{beam.feat.pvals}{A list containing feature-level p-values from compute_feature_pvalues.}

\item{set.id}{A character specifying the name of a set. Must be in beam.result$beam.data$set.data}

\item{pair.type}{A character vector. Default NULL, in which case clinical plots for all omic/endpoint pairs are produced. Otherwise specify pairs from beam.stats$beam.specs$name}

\item{number.pairs}{A numeric. Default 1, in which case only feature with best simple test for each pair is plotted. If >1, show top n simple plots ordered by feature-level p-value}

\item{n.col}{A numeric. Specify the number of columns for the plot layout; default NULL will use the number of omics types.}

\item{n.row}{A numeric. Specify the number of rows for the plot layout; default NULL will automatically define the number of rows after number of columns specified.}

\item{title.size}{A numeric. Specify the size of individual plot titles. Default is 10.}
}
\value{
A figure (ggarrange object)
}
\description{
plot_beam_clin produces a matrix of feature level clinical plots for a set. Users can specify which omic/endpoint pairs they want to see as well as the number of features from the set. Default is all omic/endpoint pairs and the top feature (smallest feature-level p-value).
}
\examples{
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=10, seed=123)
specs <- prep_beam_specs(beam.data=test.beam.data, endpts=c("MRD29", "EFS"), firth=TRUE)
test.beam.stats <- compute_beam_stats(beam.data=test.beam.data, beam.specs=specs)
test.pvals <- compute_set_pvalues(beam.stats=test.beam.stats)
test.feat.pvals <- compute_feature_pvalues(beam.stats=test.beam.stats)
plot.specs <- prep_beam_plot(beam.data=test.beam.data, beam.specs=specs)
test.plot <- plot_beam_clin(beam.result=test.beam.stats, beam.specs=plot.specs,
                            beam.set.pvals=test.pvals, beam.feat.pvals=test.feat.pvals,
                            set.id="ENSG00000099810", pair.type=NULL, number.pairs=1,
                            n.col=4, n.row=NULL, title.size=11)
}
