% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_beam_specs.R
\name{prep_beam_specs}
\alias{prep_beam_specs}
\title{Prepare beam.specs data.frame for BEAM model fitting}
\usage{
prep_beam_specs(beam.data, endpts, adjvars = NULL)
}
\arguments{
\item{beam.data}{A beam.data object from prep_beam_data}

\item{endpts}{A vector of endpoint variable names in main.data}

\item{adjvars}{Default NULL, optional vector of adjustment variable names in main.data}
}
\value{
beam.specs, a data.frame specifying the omics-endpoint association models to be fit
}
\description{
Prepare beam.specs data.frame for BEAM model fitting
}
\examples{
data(clinf)
data(omicdat)
data(omicann)
data(setdat)
data(specs)
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=100, seed=123)
#Without adjustment
prep_beam_specs(beam.data=test.beam.data, endpts=c("MRD29", "OS", "EFS"))
# With adjustment
prep_beam_specs(beam.data=test.beam.data, endpts=c("OS", "EFS"), adjvars=c("MRD29"))
}
