% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_beam_data.R
\name{print_beam_data}
\alias{print_beam_data}
\title{Print summary information about a beam.data object}
\usage{
print_beam_data(beam.data)
}
\arguments{
\item{beam.data}{A beam.data object from prep_beam_data()}
}
\value{
Messages about the beam.data object
}
\description{
Print summary information about a beam.data object
}
\examples{
data(clinf)
data(omicdat)
data(omicann)
data(setdat)
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=100, seed=123)
print_beam_data(test.beam.data)
}
