% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAM-package.R
\docType{package}
\name{BEAM-package}
\alias{BEAM}
\alias{BEAM-package}
\title{BEAM: Bootstrap Evaluation of Association Matrices}
\description{
Integrate multiple forms of high dimensional genomic data with multiple clinical endpoints to identify meaningful genes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/annaSeffernick/BEAM}
  \item \url{https://annaseffernick.github.io/BEAM/}
  \item Report bugs at \url{https://github.com/annaSeffernick/BEAM/issues}
}

}
\author{
\strong{Maintainer}: Anna Eames Seffernick \email{anna.seffernick@stjude.org} (\href{https://orcid.org/0000-0003-0848-4604}{ORCID})

Authors:
\itemize{
  \item Stanley Pounds \email{stanley.pounds@stjude.org}
  \item Xueyuan Cao \email{xcao12@uthsc.edu}
}

}
\keyword{internal}
